requirements.txt
quart
quart-cors
requests
json
python-dotenv